chore: move isTopicGroup to topicUtils and relocate test helper functions
