+++ src/utils/topicUtils.ts
@@ -1,14 +1,18 @@
-import { TopicObject, topics } from '@/data/topics';
+import { TopicHead, TopicObject, topics, TopicsGroup } from '@/data/topics';
+
+export function isTopicGroup(topicHead: TopicHead): topicHead is TopicsGroup {
+  return Array.isArray(Object.values(topicHead)[0]);
+}
 
 export function matchPath(pathname: string): boolean {
   const normalizedPath = pathname.endsWith('/') ? pathname.slice(0, -1) : pathname;
-  console.log(pathname)
   if(normalizedPath === '/search') return true;
   return topics.some((topicHead) => {
-    if('name' in topicHead && typeof topicHead.name === 'string') return ('/' + topicHead.name) === normalizedPath;
-    
-    const topicCategory = Object.values(topicHead)[0];
-    return topicCategory.some((topic: TopicObject) => ('/' + topic.name) === normalizedPath);
+    if(isTopicGroup(topicHead)){
+      const topicCategory = Object.values(topicHead)[0];
+      return topicCategory.some((topic: TopicObject) => ('/' + topic.name) === normalizedPath);
+    }
+    return ('/' + topicHead.name) === normalizedPath;
   });
 }
 
@@ -18,13 +22,12 @@ export function humanizePath(pathname: string){
 
 export function searchForEndpoint(topicName: string): string{
   const topicEndpoint = topics.map(topicHead => {
-    if ('name' in topicHead && 'endpoint' in topicHead &&  typeof topicHead.name === 'string' && typeof topicHead.endpoint === 'string') {
-      return topicHead.name === topicName ? topicHead.endpoint : undefined;
+    if(isTopicGroup(topicHead)){
+      const topicCategory: TopicObject[] = Object.values(topicHead)[0];
+      const foundTopic = topicCategory.find((topic) => topic.name === topicName);
+      return foundTopic ? foundTopic.endpoint : undefined;
     }
-
-    const topicCategory: TopicObject[] = Object.values(topicHead)[0];
-    const foundTopic = topicCategory.find((topic) => topic.name === topicName);
-    return foundTopic ? foundTopic.endpoint : undefined;
+    return topicHead.name === topicName ? topicHead.endpoint : undefined; 
   }).find(endpoint => endpoint !== undefined);
 
   if(!topicEndpoint && topicName !== 'search') throw Error('Undefined Topic');
